drivers: arm_gic: allow zero secure interrupts
Currently the generic GIC driver code has an assert to halt if no valid
pointer to the list of secure interrupts is specified.
Rework this to barf only if the number of secure interrupts is greater
than zero.
This allows to specify zero secure interrupts and pass a NULL pointer
for the table address.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>