bl31: remove no longer needed boot0 header
As ATF now lives in SRAM, we can't load it directly as the ATF binary
anymore (it would always be loaded into DRAM then).
Instead we disguise it as the SCP, which does not require a specific
boot0 header.
Remove all the code that was prefixing the binary with the boot0 header
(which was a bit misplaced in generic code anyway).

Signed-off-by: Andre Przywara <andre.przywara@arm.com>