sun50i: switch RSB and SRAM1 to be secure-only
The default security setup makes all peripherals available to the
non-secure world.
Set the RSB (which connects to the PMIC, which is exclusively controlled
by firmware) and SRAM A1 to be secure-only.
This prevents accidental (or malicious) tinkering with the PMIC, which is
potentially harmful to the board.
NOTE: This does not seem to work as expected, since access is apparently
still possile from the non-secure side.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>