sun50i: CPU ops: disable delay loops
The CPU power down/up code has delays in the sequence. It turns out that
the udelay() function has been optimized away by the compiler, so there
is no delay in between the steps in the sequence.
Moreover actually having those delays in the sequence breaks the code
(the PSCI handler hangs).
Until we know if and what delays are actually needed, lets change the
code to be easily able to turn on and off the delays - and turn it off
for now to mimic the current behaviour.
This allows fixing (and reusing) the udelay() function in a next step.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>