sunxi: add most basic H5 support
The Allwinner H5 SoC is quite close to the A64, but will mostly paired
with simple I2C voltage controllers in contrast to the complex PMIC A64
boards use these days.
Read the SoC ID and only initialize the PMIC when an A64 is detected.
Just print the SoC name if it is an H5, the regulator does not need to
be setup initially (though we may need it later).