PSCI: revert Allwinner changes
In their published ATF tree Allwinner changed some part of the
architectural PSCI code (under the services/std_svc/psci directory).
This is not only the wrong way to do, but also harmful, since they
swapped the order of "cleaning the cache" and the platform specific
"core power down" code.
With the removal of the arisc calling code, we were actually turning off
the cores in that part, so not having the caches cleaned before leads
to random data corruption and most likely crashes.
Revert that unnecessary change to this platform-agnostic code part,
also some other changes which we don't need anymore.
That fixes crashes when offlining cores, but still hangs the firmware
when the cores come back, because some locks are never released.
This will be fixed in the next commits.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>