PSCI: add core_power_down_wfi() hook for late power-off
Without a management processor to pull a core's "plug" remotely, we need
to turn our own power off at some point. Doing this in the middle of
the PSCI handling code during the platform-specific CPU_OFF callback is
a bad idea, as some locks are held at this point. As the core stops executing
instructions immediately when power gated, we never get a chance to
release those locks.
In the style of mainline ATF, introduce a callback function to do some
late platform-specific power-off work. This callback is called as the
last action in the process, just before sending the core to WFI.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>