sunxi: fix udelay()
The author of this "improved" udelay() inline assembly implementation
should be smacked for his stupidity: not specifying a _modified_ input
operand also as an output operand is asking for trouble, as the compiler
still expects the old value in the register. In this case GCC was
reusing the register for the next udelay() call, although the register
has been counted down to zero before.
Let the compiler know that this register changes, and force it to reload
the value shall it need it again.
This fixes udelay() and allows it to use the real values, so we can do
away with our hack to turn the udelays off.

Fixes: 423bc525f51a and 4409b3959edf.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>