sunxi: power: remove PMIC register molly guards
So far we were checking some PMIC registers for sensible values, to
avoid overwriting deliberately setup voltages.
However this does not work well with power control registers, as
nowadays the kernel may enable additional rails, so that a reboot will
lead to having other rails enabled, with ATF bailing out then.
Remove the comparison for the DCDC1 voltage and the power on-off control
2 register, simplifying the code on the way.